refactor(sdk): extract shared env and HTTP helpers to internal/
Move readModName, readEnginePort, readRpcPort into internal/env.ts and
readRawBody/readJsonBody into internal/http.ts so mcp.ts (Task 20) can
reuse them without duplication.
